+++ confluence-client/src/main/kotlin/com/github/zeldigas/confclient/model/User.kt
@@ -1,0 +1,3 @@
+package com.github.zeldigas.confclient.model
+
+data class User(val type: String?, val username: String?, val userKey: String?, val displayName: String?)